refactor: MenuGroup 패키지 변경
