+++ src/main/java/kitchenpos/menugroup/domain/JpaMenuGroupDao.java
@@ -1,6 +1,5 @@
-package kitchenpos.dao;
+package kitchenpos.menugroup.domain;
 
-import kitchenpos.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaMenuGroupDao extends JpaRepository<MenuGroup, Long>, MenuGroupDao {
