+++ src/main/java/kitchenpos/menugroup/domain/MenuGroupDao.java
@@ -1,6 +1,4 @@
-package kitchenpos.dao;
-
-import kitchenpos.domain.MenuGroup;
+package kitchenpos.menugroup.domain;
 
 import java.util.List;
