+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -14,12 +14,12 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.menu.domain.MenuDao;
-import kitchenpos.dao.MenuGroupDao;
+import kitchenpos.menugroup.domain.MenuGroupDao;
 import kitchenpos.dao.OrderDao;
 import kitchenpos.dao.OrderTableDao;
 import kitchenpos.product.domain.ProductDao;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderStatus;
